move failure definitions to higher level
