+++ src/Swarm/Game/Achievement/Persistence.hs
@@ -7,9 +7,9 @@ import Control.Carrier.Lift (sendIO)
 import Control.Monad (forM, forM_)
 import Data.Either (partitionEithers)
 import Data.Yaml qualified as Y
+import Swarm.Failure
 import Swarm.Game.Achievement.Attainment
 import Swarm.Game.Achievement.Definitions
-import Swarm.TUI.Model.Failure
 import Swarm.Util.GameData (getSwarmXdgDataSubdir)
 import System.Directory (
   doesDirectoryExist,
