+++ src/Swarm/Game/ScenarioInfo.hs
@@ -66,9 +66,9 @@ import Data.Text (Text)
 import Data.Time (NominalDiffTime, ZonedTime, diffUTCTime, zonedTimeToUTC)
 import Data.Yaml as Y
 import GHC.Generics (Generic)
+import Swarm.Failure
 import Swarm.Game.Entity
 import Swarm.Game.Scenario
-import Swarm.TUI.Model.Failure
 import Swarm.Util.GameData (getDataDirSafe, getSwarmSavePath)
 import System.Directory (canonicalizePath, doesDirectoryExist, doesFileExist, listDirectory)
 import System.FilePath (pathSeparator, splitDirectories, takeBaseName, takeExtensions, (-<.>), (</>))
