+++ src/Swarm/TUI/Model/UI.hs
@@ -58,6 +58,7 @@ import Data.Map (Map)
 import Data.Map qualified as M
 import Data.Text (Text)
 import Data.Text qualified as T
+import Swarm.Failure (SystemFailure)
 import Swarm.Game.Achievement.Attainment
 import Swarm.Game.Achievement.Definitions
 import Swarm.Game.Achievement.Persistence
@@ -67,7 +68,6 @@ import Swarm.Game.ScenarioInfo (
 import Swarm.Game.World qualified as W
 import Swarm.TUI.Attr (swarmAttrMap)
 import Swarm.TUI.Inventory.Sorting
-import Swarm.TUI.Model.Failure (SystemFailure)
 import Swarm.TUI.Model.FailureRender (prettyFailure)
 import Swarm.TUI.Model.Goal
 import Swarm.TUI.Model.Menu
