+++ swarm.cabal
@@ -85,6 +85,7 @@ library
                       Data.BoolExpr.Simplify
                       Swarm.App
                       Swarm.DocGen
+                      Swarm.Failure
                       Swarm.Game.Achievement.Attainment
                       Swarm.Game.Achievement.Definitions
                       Swarm.Game.Achievement.Description
@@ -138,7 +139,6 @@ library
                       Swarm.TUI.Inventory.Sorting
                       Swarm.TUI.List
                       Swarm.TUI.Model
-                      Swarm.TUI.Model.Failure
                       Swarm.TUI.Model.FailureRender
                       Swarm.TUI.Model.Goal
                       Swarm.TUI.Model.Menu
